Move ccdb access task to tutorials
